Noramlized dialogs

+++ packages/web/web-app/src/app/auth/auth-web-logout.button.tsx
@@ -1,17 +1,22 @@
 import LogoutIcon from '@mui/icons-material/Logout'
 import { Button, Grid, Typography } from '@mui/material'
-import type React from 'react'
+import React from 'react'
+import { createPortal } from 'react-dom'
 import { useNavigate } from 'react-router'
 import { toast } from 'react-toastify'
 
+import { sleep } from '@dx3/utils-shared'
 import { logger } from '@dx3/web-libs/logger'
 import { ConfirmationDialog } from '@dx3/web-libs/ui/dialog/confirmation.dialog'
+import { CustomDialog } from '@dx3/web-libs/ui/dialog/dialog.component'
+import { MODAL_ROOT_ELEM_ID } from '@dx3/web-libs/ui/system/ui.consts'
 
 import { WebConfigService } from '../config/config-web.service'
 import { useStrings } from '../i18n'
-import { useAppDispatch } from '../store/store-web-redux.hooks'
+import { useAppDispatch, useAppSelector } from '../store/store-web-redux.hooks'
 import { StyledAccountMenuListItem } from '../ui/menus/app-menu-account.ui'
 import { uiActions } from '../ui/store/ui-web.reducer'
+import { selectIsMobileWidth } from '../ui/store/ui-web.selector'
 import { useLogoutMutation } from './auth-web.api'
 import { authActions } from './auth-web.reducer'
 
@@ -21,6 +26,8 @@ type LogoutButtonType = {
 }
 
 export const LogoutButton: React.FC<LogoutButtonType> = ({ context, onLocalClick }) => {
+  const [confirmOpen, setConfirmOpen] = React.useState(false)
+  const isMobileWidth = useAppSelector((state) => selectIsMobileWidth(state))
   const dispatch = useAppDispatch()
   const navigate = useNavigate()
   const [requestLogout] = useLogoutMutation()
@@ -32,44 +39,49 @@ export const LogoutButton: React.FC<LogoutButtonType> = ({ context, onLocalClick
       onLocalClick()
     }
 
-    try {
-      dispatch(
-        uiActions.appDialogSet(
-          <ConfirmationDialog
-            cancelText={strings.CANCEL}
-            noAwait={true}
-            // bodyMessage="Is it really time to go?"
-            okText={strings.LOG_OUT}
-            onComplete={async (isConfirmed: boolean) => {
-              if (isConfirmed) {
-                try {
-                  const logoutResponse = await requestLogout().unwrap()
-                  if (logoutResponse.loggedOut) {
-                    dispatch(authActions.tokenRemoved())
-                    dispatch(authActions.setLogoutResponse(true))
-                    // toast.success('Logged out.');
-                    navigate(ROUTES.AUTH.LOGIN)
-                    setTimeout(() => dispatch(uiActions.appDialogSet(null)), 1500)
-                    return
-                  }
-                  setTimeout(() => dispatch(uiActions.appDialogSet(null)), 1500)
-                } catch (err) {
-                  logger.error(err)
-                  toast.warn(strings.COULD_NOT_LOGOUT)
-                }
-              }
-
-              if (!isConfirmed) {
-                dispatch(uiActions.appDialogSet(null))
-              }
-            }}
-          />,
-        ),
-      )
-    } catch (err) {
-      logger.error(err)
-    }
+    setConfirmOpen(true)
   }
+
+  const confirmModal = createPortal(
+    <CustomDialog
+      body={
+        <ConfirmationDialog
+          bodyMessage={''}
+          cancelText={strings.CANCEL}
+          okText={strings.LOG_OUT}
+          onComplete={async (isConfirmed: boolean) => {
+            if (isConfirmed) {
+              try {
+                const logoutResponse = await requestLogout().unwrap()
+                if (logoutResponse.loggedOut) {
+                  navigate(ROUTES.AUTH.LOGIN)
+                  dispatch(uiActions.toggleMenuSet(false))
+                  sleep(500).then(() => {
+                    setConfirmOpen(false)
+                    sleep(500).then(() => {
+                      dispatch(authActions.tokenRemoved())
+                      dispatch(authActions.setLogoutResponse(true))
+                    })
+                  })
+                }
+              } catch (err) {
+                logger.error(err)
+                toast.warn(strings.COULD_NOT_LOGOUT)
+              }
+            }
+
+            if (!isConfirmed) {
+              setConfirmOpen(false)
+            }
+          }}
+        />
+      }
+      closeDialog={() => setConfirmOpen(false)}
+      isMobileWidth={isMobileWidth}
+      open={confirmOpen}
+    />,
+    document.getElementById(MODAL_ROOT_ELEM_ID) as HTMLElement,
+  )
 
   if (context === 'APP_BAR') {
     return (
@@ -87,20 +99,29 @@ export const LogoutButton: React.FC<LogoutButtonType> = ({ context, onLocalClick
             <Typography variant="body2">{strings.LOGOUT}</Typography>
           </Grid>
         </Grid>
+        {confirmModal}
       </StyledAccountMenuListItem>
     )
   }
 
   if (context === 'APP_MENU') {
     return (
-      <Button
-        endIcon={<LogoutIcon />}
-        onClick={logout}
-      >
-        {strings.LOGOUT}
-      </Button>
+      <>
+        <Button
+          endIcon={<LogoutIcon />}
+          onClick={logout}
+        >
+          {strings.LOGOUT}
+        </Button>
+        {confirmModal}
+      </>
     )
   }
 
-  return <Button onClick={logout}>{strings.LOGOUT}</Button>
+  return (
+    <>
+      <Button onClick={logout}>{strings.LOGOUT}</Button>
+      {confirmModal}
+    </>
+  )
 }
